Preparation for switching the KIR ownership from Fusion to Kernel
